PP-10543: fix inconsistent test results by separating test files
Move the status consistency checker into its own integration test.
The main delivery queue test requires background processes to be running
which seems to cause a deadlock when inserting each status enum.
