+++ src/test/java/uk/gov/pay/webhooks/deliveryqueue/WebhookDeliveryQueueIT.java
@@ -3,13 +3,10 @@ package uk.gov.pay.webhooks.deliveryqueue;
 import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
-import io.dropwizard.testing.ConfigOverride;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 import uk.gov.pay.extension.AppWithPostgresAndSqsExtension;
 import uk.gov.pay.rule.SqsTestDocker;
 import uk.gov.pay.webhooks.ledger.LedgerStub;
@@ -22,14 +19,13 @@ import java.util.Map;
 import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.exactly;
 import static com.github.tomakehurst.wiremock.client.WireMock.matchingJsonPath;
-import static com.github.tomakehurst.wiremock.client.WireMock.notFound;
 import static com.github.tomakehurst.wiremock.client.WireMock.post;
 import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
+import static io.dropwizard.testing.ConfigOverride.config;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static uk.gov.pay.webhooks.ledger.TransactionFromLedgerFixture.aTransactionFromLedgerFixture;
 import static uk.gov.pay.webhooks.util.SNSToSQSEventFixture.anSNSToSQSEventFixture;
 
@@ -37,11 +33,7 @@ import static uk.gov.pay.webhooks.util.SNSToSQSEventFixture.anSNSToSQSEventFixtu
 public class WebhookDeliveryQueueIT {
     private static final int webhookCallbackEndpointStubPort = PortFactory.findFreePort();
     @RegisterExtension
-    public static AppWithPostgresAndSqsExtension app = new AppWithPostgresAndSqsExtension(
-            ConfigOverride.config("queueMessageReceiverConfig.backgroundProcessingEnabled", "true"),
-            ConfigOverride.config("webhookMessageSendingQueueProcessorConfig.initialDelayInMilliseconds", "0"),
-            ConfigOverride.config("webhookMessageSendingQueueProcessorConfig.threadDelayInMilliseconds", "10")
-    );
+    public static AppWithPostgresAndSqsExtension app = new AppWithPostgresAndSqsExtension(config("queueMessageReceiverConfig.backgroundProcessingEnabled", "true"));
     @RegisterExtension
     public static WireMockExtension wireMock = WireMockExtension.newInstance()
             .options(wireMockConfig().port(app.getWireMockPort()).httpsPort(webhookCallbackEndpointStubPort))
@@ -55,19 +47,11 @@ public class WebhookDeliveryQueueIT {
         dbHelper.truncateAllData();
     }
 
-    @ParameterizedTest
-    @EnumSource(value = DeliveryStatus.class)
-    public void deliveryStatusEnumIsConsistentWithDatabase(DeliveryStatus status) {
-        app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhooks VALUES (1, '2022-01-01', 'webhook-external-id', 'signing-key', 'service-id', true, 'https://callback-url.test', 'description', 'ACTIVE')"));
-        app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhook_messages VALUES (1, 'message-external-id', '2022-01-01', 1, '2022-01-01', 1, '{}', 'transaction-external-id', 'payment')"));
-        assertDoesNotThrow(() -> app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhook_delivery_queue VALUES (1, '2022-01-01', '2022-01-01', '200', 200, 1, '%s', 1250)".formatted(status))));
-    }
-
     @Test
     public void webhookMessageIsEmittedForSubscribedWebhook() throws IOException, InterruptedException {
         var serviceExternalId = "a-valid-service-id";
         dbHelper.addWebhookWithSubscription("a-valid-webhook-id", serviceExternalId, "http://localhost:%d/a-test-endpoint".formatted(app.getWireMockPort()));
-        
+
         var transaction = aTransactionFromLedgerFixture();
         var sqsMessage = anSNSToSQSEventFixture()
                 .withBody(Map.of(
@@ -90,10 +74,10 @@ public class WebhookDeliveryQueueIT {
                 exactly(1),
                 postRequestedFor(urlEqualTo("/a-test-endpoint")).withRequestBody(matchingJsonPath("$.resource_id", equalTo(transaction.getTransactionId())))
         );
-        
+
         given().port(app.getAppRule().getLocalPort())
                 .contentType(ContentType.JSON)
-                .get("/v1/webhook/webhook-external-id/message")
+                .get("/v1/webhook/a-valid-webhook-id/message")
                 .then()
                 .body("results[0].latest_attempt.status", is("SUCCESSFUL"))
                 .body("results[0].last_delivery_status", is("SUCCESSFUL"));
