+++ src/test/java/uk/gov/pay/webhooks/deliveryqueue/WebhookDeliveryQueueStatusIT.java
@@ -1,0 +1,31 @@
+package uk.gov.pay.webhooks.deliveryqueue;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.RegisterExtension;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+import uk.gov.pay.extension.AppWithPostgresAndSqsExtension;
+import uk.gov.pay.webhooks.util.DatabaseTestHelper;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+
+public class WebhookDeliveryQueueStatusIT {
+    @RegisterExtension
+    public static AppWithPostgresAndSqsExtension app = new AppWithPostgresAndSqsExtension();
+    private DatabaseTestHelper dbHelper;
+
+    @BeforeEach
+    public void setUp() {
+        dbHelper = DatabaseTestHelper.aDatabaseTestHelper(app.getJdbi());
+        dbHelper.truncateAllData();
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = DeliveryStatus.class)
+    public void deliveryStatusEnumIsConsistentWithDatabase(DeliveryStatus status) {
+        app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhooks VALUES (1, '2022-01-01', 'webhook-external-id', 'signing-key', 'service-id', true, 'https://callback-url.test', 'description', 'ACTIVE')"));
+        app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhook_messages VALUES (1, 'message-external-id', '2022-01-01', 1, '2022-01-01', 1, '{}', 'transaction-external-id', 'payment')"));
+        assertDoesNotThrow(() -> app.getJdbi().withHandle(h -> h.execute("INSERT INTO webhook_delivery_queue VALUES (1, '2022-01-01', '2022-01-01', '200', 200, 1, '%s', 1250)".formatted(status))));
+    }
+}
